+++ cardano-cli/src/Cardano/CLI/Shelley/Parsers.hs
@@ -718,6 +718,7 @@ pTransaction =
             <*> many pMetadataFile
             <*> optional pProtocolParamsSourceSpec
             <*> optional pUpdateProposalFile
+            <*  pOutputSerialisation -- TODO delete in the next release after 1.35.4
             <*> (OutputTxBodyOnly <$> pTxBodyFile Output <|> pCalculatePlutusScriptCost)
 
   pChangeAddress :: Parser TxOutChangeAddress
@@ -754,6 +755,7 @@ pTransaction =
                <*> many pMetadataFile
                <*> optional pProtocolParamsSourceSpec
                <*> optional pUpdateProposalFile
+               <*  pOutputSerialisation -- TODO delete in the next release after 1.35.4
                <*> pTxBodyFile Output
 
   pTransactionSign  :: Parser TransactionCmd
@@ -1752,6 +1754,14 @@ pOutputFormat =
     <> Opt.help "Optional output format. Accepted output formats are \"hex\" \
                 \and \"bech32\" (default is \"bech32\")."
     <> Opt.value OutputFormatBech32
+    )
+
+pOutputSerialisation :: Parser ()
+pOutputSerialisation =
+  Opt.flag () ()
+    (  Opt.long "cddl-format"
+    <> Opt.help "Serialise in the ledger CDDL specified CBOR format."
+    <> Opt.hidden
     )
 
 pMaybeOutputFile :: Parser (Maybe OutputFile)