+++ cardano-cli/test/Test/Golden/Shelley/TextEnvelope/Tx/Tx.hs
@@ -35,6 +35,7 @@ golden_shelleyTx = propertyOnce . H.moduleWorkspace "tmp" $ \tempDir -> do
     , "--tx-out", "addr_test1qpmxr8d8jcl25kyz2tz9a9sxv7jxglhddyf475045y8j3zxjcg9vquzkljyfn3rasfwwlkwu7hhm59gzxmsyxf3w9dps8832xh+1199989833223"
     , "--tx-out", "addr_test1vpqgspvmh6m2m5pwangvdg499srfzre2dd96qq57nlnw6yctpasy4+10000000"
     , "--fee", "166777"
+    , "--cddl-format" -- TODO delete in the next release after 1.35.4
     , "--out-file", transactionBodyFile
     ]