+++ cardano-cli/ChangeLog.md
@@ -7,7 +7,9 @@
 - Allow reading and writing text envelope files from pipes([PR4384](https://github.com/input-output-hk/cardano-node/pull/4384))
 - Update cardano transaction view command to render inline datums and reference inputs ([PR4089](https://github.com/input-output-hk/cardano-node/pull/4089))
 - Switch default era in cardano-cli from Alonzo to Babbage ([PR4485](https://github.com/input-output-hk/cardano-node/pull/4485))
-- Default to the ledger's CDDL format for transaction body creation by removing flags `--cddl-format` and `--cli-format` from `build` and `build-raw` ([PR 4303](https://github.com/input-output-hk/cardano-node/pull/4303))
+- For the `build` and `build-raw` commands, always use ledger's CDDL format for transaction body creation.
+  Remove the `cli-format` flag and make `--cddl-format` deprecated and hidden
+  ([PR 4303](https://github.com/input-output-hk/cardano-node/pull/4303) [PR 4617](https://github.com/input-output-hk/cardano-node/pull/4617))
 - Allow reading signing keys from a pipe ([PR 4342](https://github.com/input-output-hk/cardano-node/pull/4342))
 
 ## 1.35.3 -- August 2022